Fix: better PCAP support for multi-channel (#105)
* [all] new 'pcap' argument with selectable pcap frame type (wpan, wpan-tap, or off); dispatcher bugfix in hanging Autogo goroutine; refactor ChannelId to uint8.

+++ types/types.go
@@ -32,7 +32,7 @@ import (
 )
 
 type NodeId = int
-type ChannelId = int
+type ChannelId = uint8
 
 const (
 	InvalidNodeId         NodeId = 0
